requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Deprecated==1.2.18
fastapi==0.115.12
fastapi-cache2==0.2.2
h11==0.16.0
idna==3.10
limits==5.2.0
packaging==25.0
pendulum==3.1.0
platformdirs==4.3.8
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-memcached==1.62
redis==4.6.0
requests==2.32.3
requests-cache==1.2.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
